requirements.txt
matplotlib
pydeck
plotly
geopandas
streamlit_authenticator
st-pages